Move close fh after warn for correct error strings
Reading the exit and error string values should be immediately
after the syscall we're wanting values for, else risks another
syscall setting them to something else.

In the case of _run_command we want to read $! and $? for the
command being run through IPC run, not the values from close.