Remove extension support from MDA
I believe this was intended to act like other MDAs which set an
EXTENSION environment variable from the address for
convenience. However, it was never documented, and the implementation
has bugs that mean it never worked.

Note that in `last if ($recipient = $self->recipient)`, the condition
uses the assignment operator. This means if $self->recipient was any
truthy value, the loop would immediately abort. Otherwise, $recipient
would get to a falsy value (probably either the empty string or undef),
and there's no way we could get past the later check
`next unless (defined $extension && ( $extension =~ /\S/ ))`.

I can guess at a few ways this *should* work, but it's not clear which
one was intended. Just remove it for now until we have a real use case.