Expand pod for bin/wsgetmail
Expand documentation of arguments and the configuration file format. Add
documentation with Microsoft links about how to set up the client
application.

We expect most wsgetmail users want to run the command line tool. With
that, we are working to make the tool pod complete documentation for
them to get up and running. This is nicer for those users, and frees up
space for the library pod to become more of a developer-oriented
reference.