Generate README.md from bin/wsgetmail
We use the script as the README source, rather than App::wsgetmail,
because we expect it to be the primary tool most users are interested
in.

We generate Markdown because it'll look nicer in GitHub; and Pod::Text
breaks the documentation links in the pod, with no option to prevent
that.