Move script to bin/wsgetmail
If the path passed to install_script() isn't available in the source
code, then CPAN can't find the script to render its pod, etc. This
commit removes the .in suffix to fix that. Committers need to know
not to commit local changes to the shebang line, but that's a price
worth paying to make a nicer experience for users and readers.