Fix default values for attributes
Moo runs the default sub for an attribute if it's not set at build
time.  In the example of command_timeout, the default sub should
be setting it to 30 if it's not defined in the config.  If it is
defined in the config, it should use that instead.

As it's written, if an attribute (command_timeout) doesn't exist
in the config, BUILDARGS returns the key with an undef value, which
satisfies the attribute check, which means the default sub is never
run and the default value never set.

This commit fixes the BUILDARGS logic to first check if the value
is defined in the config before setting it into the attribute.