Add file locking to prevent contention
This change adds file locking based on the username and folder from the
merged configuration (config file + comamnd-line overrides).

This prevents multiple isntances running simultaneously on the same
username/folder while still allowing multiple instances to run, even
when they use the same base configuration.

The Microsoft Graph API is based on REST, which is stateless so there's
nothing at the remote endpoint that will prevent multiple instances of
an application from working in parallel.

In order to reduce the number of remote data fetches, wsgetmail executes
a search and pulls down a worklist of messages before it starts
processing the first message.  This decision means that multiple copies
of wsgetmail working on the same username/folder can have overlapping
sets of messages, leading to messages being processed multiple times.