Add config option strip_cr
This option directs wsgetmail to convert CRLF line endings in the
fetched message to LF line endings.  The default value is to preserve
CRLFs in fetched messages.

This behavior is tenchnically non-standard according to the MIME and
SMTP standards, but is the default behavior for at least one other mail
fetching utility and some *nix programs commonly used for mail
processing don't properly handle messages with CRLFs.