Make failure to fetch folders fatal
Prior to this change, a failure to fetch the folder details was treated
as a minor error, even though it completely prevents wsgetmail from
fetching and processing messages.

With this change wsgetmail will die with an error message and exit with
a non-zero return code if there is an issue fetching the folder details.