Count and report on failures in message processing.
Prior to this change wsgetmail reported only on the messages that were
sucessfullly processed.  Messages that resulted in errors during
post-fetch processing were not counted or reported on in any way.

With this change a count is kept of messages that resulted in errors
during post-fetch processing and this count is reported if non-zero.

wsgetmail now exits with a failure code if there were any errors
processing messages.