Add response header logging with debug on failure
A customer working with MS support was requested to provide response
header information for troubleshooting an issue.  This information was
not previously logged.

This is currently being treated as a one-off patch and not necessarily a
permanent new feature.  As a new feature it may be desirable to add
debug levels or content flags.