Ignore server-side errors on some APIs
The MS API sometimes returns 5xx errors to the requests for folder
details, folder message listing, message details, and message deletion.

It's been our observation with the deletes that, at least most of the
time, the delete happens in spite of the error code response.

This generally seems to be a temporary condition that won't recur on the
next run and there's also no known action to take when it occurs, except
to retry again later.

Because there's no action needed these messages are now suppressed by
default.

The new --response-handling command-line option can be used to disable
the new behavior or to add similar handling for additional response
code/API combinations.